Organize desktop adapter
